+++ src/content_rules/conditions/content_type.ts
@@ -11,6 +11,13 @@ export const content_type = {
     req.i18n(
       'Apply only when the current content object is of a particular type',
     ),
+  getSummary: (req: Request, params: any) =>
+    req.i18n('Content type: {types}', {
+      types:
+        params.check_types && params.check_types.length > 0
+          ? params.check_types.join(', ')
+          : req.i18n('None'),
+    }),
   schema: {
     fieldsets: [
       {