+++ src/content_rules/conditions/file_extension.ts
@@ -9,6 +9,10 @@ export const file_extension = {
   getTitle: (req: Request) => req.i18n('File Extension'),
   getDescription: (req: Request) =>
     req.i18n('Apply only to a particular file extension'),
+  getSummary: (req: Request, params: any) =>
+    req.i18n('File extension is {extension}', {
+      extension: params.file_extension || req.i18n('None'),
+    }),
   schema: {
     fieldsets: [
       {